Create select element
When receive data from the API create a select element with the id
of the objects that come from the API, if the data is not receive
create a div with no data.  Add a test to see if the API data is
correct and depending of this data, paint a select or a div.